Refactor frontend structure: Remove old pages and components, introduce new layout and sections for improved navigation and organization. Added About, Chat, Work, and Writings pages with respective sections. Implemented a new Layout component to manage common UI elements and state.
Most importantly: Updated routing and session storage handling for better user experience.
