+++ apps/frontend-public/src/App.jsx
@@ -1,8 +1,12 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Home from './app/pages/Home.jsx';
-import Post from './app/pages/Post.jsx';
-import Project from './app/pages/Project.jsx';
-import NotFound from './app/pages/NotFound.jsx';
+import Layout from './app/components/Layout.jsx';
+import AboutPage from './app/pages/AboutPage.jsx';
+import WorkPage from './app/pages/WorkPage.jsx';
+import WritingsPage from './app/pages/WritingsPage.jsx';
+import ChatPage from './app/pages/ChatPage.jsx';
+import Post from './app/pages/subpages/Post.jsx';
+import Project from './app/pages/subpages/Project.jsx';
+import NotFound from './app/pages/subpages/NotFound.jsx';
 import './App.css';
 
 function App() {
@@ -10,12 +14,17 @@ function App() {
     <Router>
       <div className="App">
         <div className="mx-auto lg:min-h-screen max-w-screen-xl lg:px-12 h-screen lg:h-auto">
-          <div className="w-full lg:h-full lg:flex lg:flex-row lg:gap-8">
+          <div className="w-full lg:h-full lg:flex lg:flex-row lg:gap-x-8">
             <Routes>
-              <Route path="/" element={<Home />} />
-              <Route path="/writings/:id" element={<Post />} />
-              <Route path="/projects/:id" element={<Project />} />
-              <Route path="*" element={<NotFound />} />
+              <Route element={<Layout />}>
+                <Route path="/" element={<AboutPage />} />
+                <Route path="/work" element={<WorkPage />} />
+                <Route path="/writings" element={<WritingsPage />} />
+                <Route path="/chat" element={<ChatPage />} />
+                <Route path="/writings/posts/:id" element={<Post />} />
+                <Route path="/work/projects/:id" element={<Project />} />
+                <Route path="*" element={<NotFound />} />
+              </Route>
             </Routes>
           </div>
         </div>
