+++ apps/frontend-public/src/app/components/PageTracker.jsx
@@ -1,33 +1,27 @@
 import { FaAngleRight } from "react-icons/fa";
-import { motion, AnimatePresence } from "framer-motion";
 import { useNavigate } from "react-router-dom";
 
-function PageTracker({ tabs, section, currentPage, onSectionClick }) {
+function PageTracker({ pathname }) {
   const navigate = useNavigate();
+  
+  // Parse pathname to get hierarchical breadcrumbs
+  const segments = pathname
+    .split('/')
+    .filter(segment => segment.length > 0);
+
 
   return (
-    <div className="flex flex-row text-sm gap-x-1 w-full h-full items-center">
-      
+    <div className="flex flex-row text-sm lg:text-xs xl:text-sm gap-x-1 w-full h-full items-center">
       <a href="https://izack.dev">izack.dev</a>
       
-          {tabs && (
-            <>
-              <FaAngleRight />
-              <span>{tabs}</span>
-            </>
-          )}
-          {section && (
-            <>
-              <FaAngleRight />
-              <span>{section}</span>
-            </>
-          )}
-          {currentPage && (
-            <>
-              <FaAngleRight />
-              <span className="flex-1 min-w-0 break-words">{currentPage}</span>
-            </>
-          )}
+      {segments.map((segment, index) => (
+        <span key={index} className="flex items-center gap-x-1">
+          <FaAngleRight />
+          <span className={index === segments.length - 1 ? 'flex-1 min-w-0 break-words' : ''}>
+            {segment}
+          </span>
+        </span>
+      ))}
     </div>
   );
 }
